環境依存への対策

+++ kalman/cpp/Lib/ESKF/inc/eskf_runner.hpp
@@ -16,7 +16,7 @@ class ESKFRunner {
 public:
     ESKFRunner();
     ~ESKFRunner() = default;
-    static void predict(ESKFState* s, const double* a_meas, const double* w_meas);
+    static void predict(ESKFState* s, const float* a_meas, const float* w_meas);
 private:
     static void apply_accel_z_integration(cmath_fx::Vector<3, float>& v, const cmath_fx::Vector<4, float>& q, const cmath_fx::Vector<3, float>& a_for_vel, float dt, const cmath_fx::Vector<3, float>& g, float accel_z_threshold, float accel_z_damping);
     static void apply_velocity_clipping(cmath_fx::Vector<3, float>& v, cmath_fx::Matrix<15, 15, float>& P, float max_vel);
